Apply new endpoint for Get fact details by id Update facy by id delete fact connect fact to another character

+++ src/LoreWeave.Api/Dtos/Characters/UpdateCharacterDto.cs
@@ -13,7 +13,7 @@ public record UpdateCharacterDto(
     public UpdateCharacterCommand ToCommand(Guid id, string version) => new(
         id,
         Name,
-        int.Parse(
+        ushort.Parse(
             version
                 .Replace(
                     "\"",
